Rename report sheet views directory
