+++ frontend/src/App.tsx
@@ -18,8 +18,8 @@ import { PaymentDetail } from './views/payments/PaymentDetail';
 import { PaymentOverview } from './views/payments/PaymentOverview';
 import { PhoneListView } from './views/PhoneList';
 import { Register } from './views/register/Register';
-import { ReportSheetOverview } from './views/report_sheets/ReportSheetOverview';
-import { ReportSheetUpdate } from './views/report_sheets/ReportSheetUpdate';
+import { ReportSheetOverview } from './views/expense_sheets/ReportSheetOverview';
+import { ReportSheetUpdate } from './views/expense_sheets/ReportSheetUpdate';
 import { ServiceOverview } from './views/service_overview/ServiceOverview';
 import { ServiceSpecificationsOverview } from './views/service_specification/ServiceSpecificationsOverview';
 import { UserFeedbackOverview } from './views/user_feedback_overview/UserFeedbackOverview';
